edit in company at 2017.3.16

+++ SquareNode.h
@@ -56,7 +56,7 @@ public:
 	void setScore(int score){ m_score = score; addScore(0); }
 
 	float getSpeed() { return m_speed; }
-	void setSpeed(int speed){ m_speed = speed; }
+	void setSpeed(float speed){ m_speed = speed; }
 
 	int getSquareType(){ return m_squareType; }
 	b2Body *getBody(){ return m_body; }
